pyproject.toml
[project]
name = "promptfoo-crewai-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.157.0",
    "crewai-tools[mcp]>=0.60.0",
    "fastapi[standard]>=0.116.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
